requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pyyaml>=6.0
browser-cookie3>=0.19.0
